hoihio, det funker

+++ services/setup.py
@@ -1,5 +1,9 @@
+"""
+$Id: setup.py,v 1.3 2002/06/27 13:01:13 erikgors Exp $
+"""
 import re,getopt,sys,config,database,psycopg,os
 from job import jobmap
+HEADER = '#id     active   sysname              handler    args'
 
 class Service:
 	def __init__(self,id,active,sysname,handler,args):
@@ -17,6 +21,8 @@ class Service:
 			return i.__cmp__(j)
 	def __repr__(self):
 		return "%-8s,%-8s,%-20s,%-10s,%s" % (`self.id`,`self.active`,`self.sysname`,`self.handler`,self.args)
+	def __hash__(self):
+		return self.id.__hash__()
 def parseLine(line):
 	try:
 		id,active,sysname,handler,args = eval(line.lower())
@@ -36,14 +42,7 @@ def parseLine(line):
 		msg = "should be a dict: (%s)" % args
 		raise TypeError(msg)
 	return Service(id,active,sysname,handler,args)
-def help():
-	print """ - Setup -
 
-	valid options:
-		-h shows help
-		-f file, default services.conf
-		-c file, default db.conf
-"""
 def fromFile(file):
 	new = []
 	for i in open(file).read().split('\n'):
@@ -60,6 +59,7 @@ class DB:
 		self.db = psycopg.connect("host = %s user = %s dbname = %s password = %s" % (conf["dbhost"],"manage",conf["db_nav"],conf["userpw_manage"]))
 		self.db.autocommit(1)
 		self.sysboks()
+		database.db = self.db
 	def sysboks(self):
 		s = self.query('select sysname,boksid from boks')
 		self.boks = dict(s)
@@ -68,8 +68,24 @@ class DB:
 		s.execute(querystring)
 		return s.fetchall()
 	def fromDB(self):
-		i = self.query('SELECT serviceid,sysname FROM service NATURAL JOIN boks')
-		return dict(i)
+		services = []
+
+		for i in database.getJobs(0):
+			serviceid = i.getServiceid()
+			active = (i.active and 'true') or 'false'
+			boksid = i.getBoksid()
+			for j in self.boks:
+				if self.boks[j] == boksid:
+					sysname = j
+					break
+			handler = i.getType()
+			args = i.getArgs()
+			
+			new = Service(serviceid,active,sysname,handler,args)
+			services += [new]
+		services.sort()
+		return services
+
 	def delete(self,serviceid):
 		s = self.db.cursor()
 		s.execute('DELETE FROM service WHERE serviceid = %s' % serviceid)
@@ -83,8 +99,8 @@ class DB:
 		s = self.db.cursor()
 		next = self.query("select nextval('service_serviceid_seq')")[0][0]
 		s.execute("INSERT INTO service (serviceid,boksid,active,handler) VALUES (%s,%s,%s,'%s')" % (next, self.boks[service.sysname], service.active, service.handler))
-		self.insertargs(service)
 		service.id = next
+		self.insertargs(service)
 		return next
 	def insertargs(self,service):
 		s = self.db.cursor()
@@ -96,25 +112,14 @@ def newFile(file,conf):
 	db = DB(conf)
 	db.connect()
 
-	database.db = db.db
-	jobs  = db.fromDB()
+	print 'fetching services from db'
+	services = db.fromDB()
 
-	services = []
-
-	for i in database.getJobs(0):
-		serviceid = i.getServiceid()
-		active = (i.active and 'true') or 'false'
-		sysname = jobs[serviceid]
-		handler = i.getType()
-		args = i.getArgs()
-		
-		new = Service(serviceid,active,sysname,handler,args)
-		services += [new]
-	services.sort()
-	header = '#id     active   sysname              handler    args'
-	print header
+	print 'creating ' + file
+	file = open(file,'w')
+	file.write(HEADER + '\n')
 	for i in services:
-		print i
+		file.write('%s\n' % i)
 	
 		
 def main(file,conf):
@@ -125,18 +130,16 @@ def main(file,conf):
 	new = {}
 	for i in fromFile(file):
 		new[i.id] = i
-	print new
 
 	print 'fetching services from db'
 	db.connect()
 	result = db.fromDB()
-	print result
 
 	delete = []
 	for i in result:
 		if i not in new:
 			delete += [i]
-		elif new[i].sysname != result[i]:
+		elif new[i.id].sysname != i.sysname:
 			msg = 'serviceid and sysname do not match: (%s,%s) should be (%s,%s)' % (i,new[i],i,result[i])
 			raise TypeError(msg)
 	for i in new.values():
@@ -144,10 +147,20 @@ def main(file,conf):
 			msg = 'sysname not found: (%s)' % i.sysname
 			raise TypeError(msg)
 
-	print 'to be deleted:', delete
-	for i in delete:
-		db.delete(i)
-	print 'to be added/updateded:', new
+	if delete:
+		print 'to be deleted:'
+		for i in delete:
+			print i
+		s = 0
+		while s not in ('yes','no'):
+			print '\nare you sure? (yes/no)'
+			s = raw_input()
+		if s == 'no':
+			print 'quitting'
+			sys.exit(1)
+		for i in delete:
+			db.delete(i)
+	print 'updating db'
 	for i in new.values():
 		if i.id == 'new':
 			db.insertservice(i)
@@ -155,28 +168,50 @@ def main(file,conf):
 			db.updateservice(i)
 	keys = new.keys()
 	keys.sort()
+
+	print 'creating ' + file
+	file = open(file,'w')
+	file.write(HEADER + '\n')
 	for i in keys:
-		print new[i]
-	
+		file.write('%s\n' % new[i])
+def help():
+	print """ - Setup -
+
+	valid options:
+		-h 	- shows help
+		-f file	- default services.conf
+		-c file	- default db.conf
+		-n	- generates a new service.conf from db
+		-u	- updates the db from services.conf
+"""
 if __name__=='__main__':
 	try:
-		opts, args = getopt.getopt(sys.argv[1:], 'hnf:c:')
+		opts, args = getopt.getopt(sys.argv[1:], 'hnuf:c:')
 		opts = dict(opts)
 		if '-h' in opts:
 			help()
 			sys.exit()
 		file = opts.get('-f','services.conf')
 		conf = opts.get('-c','db.conf')
+		if not opts:
+			help()
+		elif not os.path.exists(conf):
+			msg = 'cant find file: ' + conf 
+			raise IOError(msg)
 		if '-n' in opts:
+			if os.path.exists(file):
+				print 'creating backup: services.backup'
+				open('services.backup','w').write(open(file).read())
 			newFile(file,conf)
-		else:
-			for i in (file,conf):
-				if not os.path.exists(i):
-					msg = 'cant find file: ' + i
+		elif '-u' in opts:
+			if not os.path.exists(file):
+					msg = 'cant find file: ' + file
 					raise IOError(msg)
+			else:
+				print 'creating backup: services.backup'
+				open('services.backup','w').write(open(file).read())
 			main(file,conf)
 		sys.exit(0)
 	except (getopt.error):
 		pass
-	help()
 	sys.exit(2)
